requirements.txt
flask
langchain==0.0.235
openai
PyPDF2
python-dotenv
pypdf
chromadb==0.3.22
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
streamlit==1.25.0